Extract from topojson
