+++ types/topojson-simplify/tsconfig.json
@@ -19,8 +19,6 @@
     },
     "files": [
         "index.d.ts",
-        "topojson-tests.ts",
-        "test/atlas-tests.ts",
-        "test/server-tests.ts"
+        "topojson-simplify-tests.ts"
     ]
 }
