+++ types/topojson-simplify/tslint.json
@@ -1,0 +1,1 @@
+{ "extends": "dtslint/dt.json" }
